requirements.txt
pandas
requests
sqlalchemy
gspread
gspread_dataframe
google-cloud-secret-manager
google-auth-oauthlib
google-api-python-client
python-dateutil
functions-framework
tenacity
sendgrid
google-cloud-logging
flask
defusedxml
psycopg2-binary==2.9.9
flask-login==0.6.3
flask-dance[sqla]==7.1.0
PyJWT==2.8.0
cryptography==41.0.7
oauthlib==3.2.2
gunicorn==21.2.0
flask-sqlalchemy==3.1.1
cryptography==41.0.7
flask-dance[sqla]==7.1.0
flask-login==0.6.3
flask-sqlalchemy==3.1.1
gunicorn==21.2.0
oauthlib==3.2.2
PyJWT==2.8.0
pytz
werkzeug
